Add files via upload
So many changes. Redid the entire file with beautifulsoup4 instead of opening the browser with selenium. Much easier to go into the network --> find a link to a page of each set of results --> then download the images from there rather than get the js to load to get all results. Downloader is set to take the first 10 pages of results.