Add files via upload
This program inserts rows into a Excel file based on user input when the program is called. Automate the Boring Stuff had an outdated/harder way of creating the program. I just used a single line from the openpyxl documentation to insert a row and loop x number of times.