Add files via upload
This program encrypts all unencrypted PDFs in a given directory.
Program seems to work when manually opening the newly encrypted files and putting in the password, but adding in the new program checker gives an error.

Error: PdfReadWarning: PdfFileReader stream/file object is not in binary mode. It may not be read correctly.
So maybe the new PDF is not saved correcty? Will have to double check.